move do_laser to main. fmt.
